requirements.txt
pandas
python-dotenv
chromadb
openai
streamlit
pytest
